refactor App.js
